griffin/streaming/requirements.txt
kafka-python==2.1.3
psycopg2-binary==2.9.10
